requirements.txt
argcomplete==1.9.4
asn1crypto==0.24.0
awscli==1.16.103
boto==2.48.0
boto3==1.9.93
botocore==1.12.93
cachetools==3.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
crcmod==1.7
cryptography==2.5
docutils==0.14
fasteners==0.14.1
futures==2.2.0
gcs-oauth2-boto-plugin==2.2
google-api-core==1.7.0
google-apitools==0.5.26
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-reauth==0.1.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
gsutil==4.34
httplib2==0.12.0
idna==2.8
jmespath==0.7.1
mock==2.0.0
monotonic==1.5
oauth2client==4.1.3
pbr==5.1.2
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-gflags==3.1.2
pytz==2018.9
pyu2f==0.1.4
PyYAML==3.13
requests==2.21.0
retry-decorator==1.1.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
SocksiPy-branch==1.1
urllib3==1.24.1
